Modularize Feature strategies
